// src/openapi3/sdn-fabric/0.1.0/models/switch.ts
// Code generated by openapi-gen. DO NOT EDIT.
/* tslint:disable */
/* eslint-disable */
import { AdditionalPropertiesUnchTarget } from './additional-properties-unch-target';
import { SwitchAttributeList } from './switch-attribute-list';
import { SwitchManagement } from './switch-management';
import { SwitchPortList } from './switch-port-list';
import { SwitchState } from './switch-state';
import { SwitchSwitchPair } from './switch-switch-pair';
import { SwitchVlanList } from './switch-vlan-list';

/**
 * A managed device in the fabric (single)
 */
export interface Switch {
  attribute?: SwitchAttributeList;

  /**
   * long description field
   */
  description?: string;

  /**
   * display name to use in GUI or CLI
   */
  'display-name'?: string;
  management?: SwitchManagement;

  /**
   * link to switch model
   */
  'model-id': string;
  port?: SwitchPortList;

  /**
   * The role of the switch in the fabric
   */
  role: 'undefined' | 'leaf' | 'spine';
  state?: SwitchState;

  /**
   * unique identifier for the switch
   */
  'switch-id': string;
  'switch-pair'?: SwitchSwitchPair;
  vlan?: SwitchVlanList;

  [key: string]: 'undefined' | 'leaf' | 'spine' | AdditionalPropertiesUnchTarget | SwitchAttributeList | SwitchManagement | SwitchPortList | SwitchState | SwitchSwitchPair | SwitchVlanList | string | undefined;
}